+++ src/pybmodes/__init__.py
@@ -182,7 +182,7 @@ from pybmodes.options import CheckOptions, FitOptions, SolverOptions
 try:
     __version__ = version("pybmodes")
 except PackageNotFoundError:
-    __version__ = "1.10.1"
+    __version__ = "1.11.0"
 
 __all__ = [
     "CheckOptions",